requirements.txt
pyserial
gurux-dlms
